refactor: OutputInterface의 로직을 LotteryResultRenderer로 이전 (#14)
